Update of displayBadges(): include badge name tooltip
This code is able to display a tooltip on hover (and an alt text for accessibility) on the badge image when called, for example on the account page. This makes the wondering user aware of how the badge they have achieved is called. Since badge names are already stored in the database, and the display of the image doesn't change, this change is not that intrusive :)